+++ packages/components/src/draggable/index.js
@@ -1,12 +1,8 @@
 /**
  * WordPress dependencies
  */
+import { throttle } from '@wordpress/compose';
 import { useEffect, useRef } from '@wordpress/element';
-
-/**
- * External dependencies
- */
-import { throttle } from 'lodash';
 
 const dragImageClass = 'components-draggable__invisible-drag-image';
 const cloneWrapperClass = 'components-draggable__clone';
@@ -180,6 +176,7 @@ export default function Draggable( {
 
 		// Aim for 60fps (16 ms per frame) for now. We can potentially use requestAnimationFrame (raf) instead,
 		// note that browsers may throttle raf below 60fps in certain conditions.
+		// @ts-ignore
 		const throttledDragOver = throttle( over, 16 );
 
 		ownerDocument.addEventListener( 'dragover', throttledDragOver );